requirements.txt
discord.py-self==1.9.2
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
python-dotenv==1.0.0
aiohttp==3.9.1
jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
setuptools>=65.5.1
wheel>=0.38.0 